V 1.3 I guess
- *BURP*
- Global data loading in comparison almost implemented
(global data value gets lost after more than 6 selections)
- Global data names get scrambled after 5 or more selections
-Tested only with dummy data

+++ TaskStatistics/src/View/UserView.java
@@ -114,7 +114,7 @@ public class UserView {
             } catch (IndexOutOfBoundsException exception) {
                 getDataNameProperties().add(new SimpleStringProperty());
                 getDataNameProperties().get(i).set("EMPTY");
-                getDataValueProperties().add(new SimpleDoubleProperty());
+                getDataValueProperties().add(new SimpleFloatProperty());
                 getDataValueProperties().get(i).setValue(0);
             }
         }
@@ -144,4 +144,6 @@ public class UserView {
     public void Update() {
         model.getMoreData();
     }
+
+
 }
